requirements.txt
fastapi
uvicorn
opencv-python
mediapipe==0.10.9
numpy
pillow
python-multipart
pyjwt